% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_analysis}
\alias{create_analysis}
\title{Create a Distill themed RMarkdown analysis}
\usage{
create_analysis(title, ..., open = interactive())
}
\arguments{
\item{title}{Title of the analysis.}

\item{...}{Arguments passed to \code{\link[distill:create_post]{distill::create_post()}}.}

\item{open}{Open the post in an editor after creating it.}
}
\description{
Create a Distill themed RMarkdown analysis
}
