% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_bhhi_site}
\alias{create_bhhi_site}
\title{Create BHHI project site}
\usage{
create_bhhi_site(dir, title, gh_pages = TRUE, open = TRUE)
}
\arguments{
\item{dir}{Path to the directory holding the site}

\item{title}{Title of the site}

\item{gh_pages}{Will the site be hosted on GitHub pages}

\item{open}{Open a new RStudio session in the site directory}
}
\description{
Create BHHI project site
}
